refactor(core): limit global imports in key places
so that by importing `apps.base`, we don't pull in the whole circus
