+++ core/src/apps/common/mnemonic.py
@@ -1,8 +1,6 @@
 import storage.device
 from trezor import ui, utils, workflow
-from trezor.crypto import bip39, slip39
 from trezor.messages import BackupType
-from trezor.ui.components.tt.text import Text
 
 if False:
     from trezor.messages.ResetDevice import EnumTypeBackupType
@@ -39,9 +37,13 @@ def get_seed(passphrase: str = "", progress_bar: bool = True) -> bytes:
         render_func = _render_progress
 
     if is_bip39():
+        from trezor.crypto import bip39
+
         seed = bip39.seed(mnemonic_secret.decode(), passphrase, render_func)
 
     else:  # SLIP-39
+        from trezor.crypto import slip39
+
         identifier = storage.device.get_slip39_identifier()
         iteration_exponent = storage.device.get_slip39_iteration_exponent()
         if identifier is None or iteration_exponent is None:
@@ -55,6 +57,8 @@ def get_seed(passphrase: str = "", progress_bar: bool = True) -> bytes:
 
 
 def _start_progress() -> None:
+    from trezor.ui.components.tt.text import Text
+
     # Because we are drawing to the screen manually, without a layout, we
     # should make sure that no other layout is running.
     workflow.close_others()
