+++ core/src/apps/common/passphrase.py
@@ -3,10 +3,7 @@ from micropython import const
 import storage.device
 from trezor import wire, workflow
 from trezor.messages import ButtonRequestType
-from trezor.messages.PassphraseAck import PassphraseAck
-from trezor.messages.PassphraseRequest import PassphraseRequest
 from trezor.ui import ICON_CONFIG, draw_simple
-from trezor.ui.components.tt.passphrase import CANCELLED, PassphraseKeyboard
 from trezor.ui.components.tt.text import Text
 
 from . import button_request
@@ -41,6 +38,9 @@ async def _request_from_user(ctx: wire.Context) -> str:
 
 
 async def _request_on_host(ctx: wire.Context) -> str:
+    from trezor.messages.PassphraseAck import PassphraseAck
+    from trezor.messages.PassphraseRequest import PassphraseRequest
+
     _entry_dialog()
 
     request = PassphraseRequest()
@@ -74,6 +74,8 @@ async def _request_on_host(ctx: wire.Context) -> str:
 
 
 async def _request_on_device(ctx: wire.Context) -> str:
+    from trezor.ui.components.tt.passphrase import CANCELLED, PassphraseKeyboard
+
     await button_request(ctx, code=ButtonRequestType.PassphraseEntry)
 
     keyboard = PassphraseKeyboard("Enter passphrase", _MAX_PASSPHRASE_LEN)
