+++ core/src/storage/device.py
@@ -2,8 +2,6 @@ from micropython import const
 from ubinascii import hexlify
 
 from storage import common
-from trezor.crypto import random
-from trezor.messages import BackupType
 
 if False:
     from trezor.messages.ResetDevice import EnumTypeBackupType
@@ -36,8 +34,6 @@ _SD_SALT_AUTH_KEY          = const(0x12)  # bytes
 INITIALIZED                = const(0x13)  # bool (0x01 or empty)
 _SAFETY_CHECK_LEVEL        = const(0x14)  # int
 _EXPERIMENTAL_FEATURES     = const(0x15)  # bool (0x01 or empty)
-
-_DEFAULT_BACKUP_TYPE       = BackupType.Bip39
 
 SAFETY_CHECK_LEVEL_STRICT  : Literal[0] = const(0)
 SAFETY_CHECK_LEVEL_PROMPT  : Literal[1] = const(1)
@@ -79,6 +75,8 @@ def is_initialized() -> bool:
 
 
 def _new_device_id() -> str:
+    from trezorcrypto import random  # avoid pulling in trezor.crypto
+
     return hexlify(random.bytes(12)).decode().upper()
 
 
@@ -121,9 +119,11 @@ def get_mnemonic_secret() -> bytes | None:
 
 
 def get_backup_type() -> EnumTypeBackupType:
+    from trezor.messages import BackupType
+
     backup_type = common.get_uint8(_NAMESPACE, _BACKUP_TYPE)
     if backup_type is None:
-        backup_type = _DEFAULT_BACKUP_TYPE
+        backup_type = BackupType.Bip39
 
     if backup_type not in (
         BackupType.Bip39,
